pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchard-ml"
version = "0.1.7"
description = "Type-Safe Deep Learning Framework for Computer Vision"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Tommaso Russo", email = "tomrussobuilds@users.noreply.github.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=2.0.0",
    "pandas>=2.0.0",  # floor kept low: pandas 3.x has issues on Python 3.14
    "scikit-learn>=1.5.0",
    "Pillow>=12.1.1",
    "h5py>=3.12.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0",
    "optuna>=4.0.0",
    "plotly>=6.0.0",
    "matplotlib>=3.9.0",
    "xlsxwriter>=3.2.0",
    "openpyxl>=3.1.0",
    "tqdm>=4.66.0",
    "requests>=2.32.0",
    "psutil>=7.2.2",
    "timm>=1.0.0",
    "onnx>=1.17.0",
    "onnxscript>=0.2.0",
    "onnxruntime>=1.20.0; python_version < '3.14'",
    "typer>=0.9.0",
]

[project.urls]
Homepage = "https://github.com/tomrussobuilds/orchard-ml"
Repository = "https://github.com/tomrussobuilds/orchard-ml"
Issues = "https://github.com/tomrussobuilds/orchard-ml/issues"
Documentation = "https://tomrussobuilds.github.io/orchard-ml/"

[project.scripts]
orchard = "orchard.cli_app:app"

[project.optional-dependencies]
tracking = [
    "mlflow>=3.9.0",
]
notebooks = [
    "ipython>=9.0.0",
]
test = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.5.0",
]
dev = [
    "orchard-ml[test]",
    "black>=26.1.0",
    "ruff>=0.11.0",
    "mypy==1.19.1",
    "bandit>=1.8.0",
    "radon>=6.0.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.32.0",
    "git-cliff>=2.12.0",
    "pre-commit>=4.0.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=1.0.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.9",
]

[tool.setuptools.packages.find]
include = ["orchard*"]
exclude = ["tests*", "docs*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
filterwarnings = [
    "ignore:AMP requires GPU.*:UserWarning",
    "ignore:CUDA was explicitly requested.*:UserWarning",
    "ignore:MPS was explicitly requested.*:UserWarning",
    "ignore:Training at resolution.*on CPU.*:UserWarning",
    "ignore:max_samples.*is less than 10x num_classes.*:UserWarning",
    "ignore:`isinstance\\(treespec, LeafSpec\\)` is deprecated.*:FutureWarning",
    "ignore:Detected call of `lr_scheduler.step\\(\\)` before `optimizer.step\\(\\)`.*:UserWarning",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests - fast, isolated component tests",
    "integration: Integration tests - test component interactions",
    "slow: Slow tests - may take several minutes",
    "gpu: Tests requiring GPU hardware",
]

[tool.coverage.run]
source = ["orchard"]
omit = [
    "*/tests/*",
    "*/conftest.py",
    "*/__pycache__/*",
    "*/__init__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.bandit]
exclude_dirs = ["tests", "docs", "build", "dist"]
skips = ["B101"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
exclude = ["tests/", "docs/", "build/", "dist/"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["notebooks"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "ARG", "D"]
ignore = [
    "E501",
    # Docstring style choices (our convention: Pattern A)
    "D212",   # multi-line-summary-first-line — we use D213 (summary on second line)
    "D203",   # blank-line-before-class — conflicts with D211
    "D413",   # missing-blank-line-after-last-section — not our style
    "D412",   # blank-lines-between-header-and-content — we add them for mkdocs
    "D200",   # unnecessary-multiline-docstring — Pattern A allows expanded one-liners
    "D401",   # non-imperative-mood — too noisy ("Applies" vs "Apply")
    "D102",   # undocumented-public-method — class docstring covers these
    "D105",   # undocumented-magic-method — __repr__/__len__ are self-explanatory
    "D107",   # undocumented-public-init — class docstring covers __init__
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F841", "E501", "W293", "ARG", "D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["orchard"]